+++ packages/likec4/app/src/routes/__root.tsx
@@ -1,3 +1,10 @@
+// SPDX-License-Identifier: MIT
+//
+// Copyright (c) 2023-2026 Denis Davydkov
+// Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+//
+// Portions of this file have been modified by NVIDIA CORPORATION & AFFILIATES.
+
 import type { NonEmptyArray, ProjectId } from '@likec4/core/types'
 import { MantineProvider } from '@mantine/core'
 import { createRootRouteWithContext, Outlet, stripSearchParams } from '@tanstack/react-router'
@@ -27,6 +34,7 @@ export const Route = createRootRouteWithContext<Context>()({
         theme: undefined,
         dynamic: 'diagram',
         relationships: undefined,
+        focusOnElement: undefined,
       }),
     ],
   },