+++ packages/likec4/src/cli/mcp/index.ts
@@ -1,8 +1,9 @@
+import { configureLanguageServerLogger } from '@likec4/language-server'
 import { fromWorkspace } from '@likec4/language-services/node'
 import k from 'tinyrainbow'
 import type { Argv } from 'yargs'
 import { boxen } from '../../logger'
-import { path, useDotBin } from '../options'
+import { logLevel, path, useDotBin, verbose, verboseLogLevel } from '../options'
 
 const mcpCmd = <T>(yargs: Argv<T>) => {
   return yargs
@@ -14,6 +15,7 @@ const mcpCmd = <T>(yargs: Argv<T>) => {
         y
           .usage(`${k.bold('Usage:')} $0 mcp [path]`)
           .positional('path', path)
+          .default('path', '.', 'current directory')
           .option('stdio', {
             boolean: true,
             description: 'use stdio transport',
@@ -31,6 +33,9 @@ const mcpCmd = <T>(yargs: Argv<T>) => {
             conflicts: 'stdio',
           })
           .option('use-dot', useDotBin)
+          .option('log-level', logLevel)
+          .option('verbose', verbose)
+          .showHidden()
           .epilog(`${k.bold('Examples:')}
 ${k.green('$0 mcp')}
   ${k.gray('Start MCP with default stdio transport')}
@@ -41,8 +46,16 @@ ${k.green('$0 mcp -p 1234')}
 `),
       handler: async args => {
         if (args.http || args.port) {
+          configureLanguageServerLogger({
+            colors: k.isColorSupported,
+            logLevel: args.verbose ? verboseLogLevel : args.logLevel,
+          })
           await startHttpMcp(args.path, args.useDot, args.port)
         } else {
+          configureLanguageServerLogger({
+            useStdErr: true,
+            logLevel: args.verbose ? verboseLogLevel : args.logLevel,
+          })
           await startStdioMcp(args.path, args.useDot)
         }
       },
@@ -53,6 +66,7 @@ async function startHttpMcp(path: string, useDotBin: boolean, port = 33335) {
   await fromWorkspace(path, {
     mcp: { port },
     watch: true,
+    configureLogger: false,
     graphviz: useDotBin ? 'binary' : 'wasm',
   })
   boxen(
@@ -78,6 +92,7 @@ async function startStdioMcp(path: string, useDotBin: boolean) {
   await fromWorkspace(path, {
     mcp: 'stdio',
     watch: true,
+    configureLogger: false,
     graphviz: useDotBin ? 'binary' : 'wasm',
   })
 }