+++ packages/layouts/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @likec4/layouts
 
+## 1.51.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @likec4/core@1.51.0
+  - @likec4/log@1.51.0
+
 ## 1.50.0
 
 ### Patch Changes